add signal raise logging

+++ utils.h
@@ -30,17 +30,17 @@ inline std::filesystem::path getExeFolder(std::string_view exePath) {
 
 inline date::hh_mm_ss<std::chrono::milliseconds> delphiTimeToTimeOfDay(const std::string &time) {
   using namespace std::string_literals;
-  const auto hours = std::chrono::hours(std::atoi(time.substr(0, 2).c_str()));
-  const auto minutes = std::chrono::minutes(std::atoi(time.substr("hh:"s.size(), 2).c_str()));
-  const auto seconds = std::chrono::seconds(std::atoi(time.substr("hh:mm:"s.size(), 2).c_str()));
+  const auto hours = std::chrono::hours(static_cast<unsigned>(std::atoi(time.substr(0, 2).c_str())));
+  const auto minutes = std::chrono::minutes(static_cast<unsigned>(std::atoi(time.substr("hh:"s.size(), 2).c_str())));
+  const auto seconds = std::chrono::seconds(static_cast<unsigned>(std::atoi(time.substr("hh:mm:"s.size(), 2).c_str())));
   return date::hh_mm_ss<std::chrono::milliseconds>{hours + minutes + seconds};
 }
 
 inline date::year_month_day delphiDateToYearMonthDay(const std::string &date) {
   using namespace std::string_literals;
-  const auto day = date::day(std::atoi(date.substr(0, "dd"s.size()).c_str()));
-  const auto month = date::month(std::atoi(date.substr("dd/"s.size(), "mm"s.size()).c_str()));
-  const auto year = date::year(std::atoi(date.substr("dd/mm/"s.size(), "yyyy"s.size()).c_str()));
+  const auto day = date::day(static_cast<unsigned>(std::atoi(date.substr(0, "dd"s.size()).c_str())));
+  const auto month = date::month(static_cast<unsigned>(std::atoi(date.substr("dd/"s.size(), "mm"s.size()).c_str())));
+  const auto year = date::year(static_cast<unsigned>(std::atoi(date.substr("dd/mm/"s.size(), "yyyy"s.size()).c_str())));
   return year / month / day;
 }
